Extract JSON serializer test into own test case.
